Changed routing to allow for proper error redirection
