+++ common/v2/features/ErrorHandling/index.ts
@@ -1,1 +1,1 @@
-export { ErrorContext, ErrorProvider } from './ErrorProvider';
+export { ErrorContext, default as ErrorProvider } from './ErrorProvider';
